worker/requirements.txt
redis
yt-dlp
ffmpeg-python
google-cloud-storage
requests
google-auth